exported c functions now call rust api
